Make meta-repo a persistent flag
