+++ src/go/cmdinit/init_command.go
@@ -2,8 +2,6 @@ package cmdinit
 
 import (
 	"fmt"
-	"os"
-	"path/filepath"
 
 	"github.com/kkrull/marmot/cmd"
 	"github.com/spf13/cobra"
@@ -16,17 +14,13 @@ func NewInitCommand() *initCommand {
 
 type initCommand struct{}
 
-func (cliCmd *initCommand) RegisterWithCobra() error {
-	if cobraCmd, cobraErr := cliCmd.toCobraCommand(); cobraErr != nil {
-		return cobraErr
-	} else {
-		cmd.AddMetaRepoCommand(*cobraCmd)
-		return nil
-	}
+func (cliCmd *initCommand) RegisterWithCobra() {
+	cobraCmd := cliCmd.toCobraCommand()
+	cmd.AddMetaRepoCommand(*cobraCmd)
 }
 
-func (cliCmd *initCommand) toCobraCommand() (*cobra.Command, error) {
-	cobraCmd := &cobra.Command{
+func (cliCmd *initCommand) toCobraCommand() *cobra.Command {
+	return &cobra.Command{
 		Long: "Initialize a new Meta Repo, if none is already present.",
 		RunE: func(cobraCmd *cobra.Command, _args []string) error {
 			config := cmd.ParseFlags(cobraCmd)
@@ -41,20 +35,5 @@ func (cliCmd *initCommand) toCobraCommand() (*cobra.Command, error) {
 		},
 		Short: "Initialize a meta repo",
 		Use:   "init",
-	}
-
-	if defaultPath, pathErr := defaultMetaRepoPath(); pathErr != nil {
-		return nil, pathErr
-	} else {
-		cobraCmd.Flags().String("meta-repo", defaultPath, "Meta repo to use")
-		return cobraCmd, nil
-	}
-}
-
-func defaultMetaRepoPath() (string, error) {
-	if homeDir, homeErr := os.UserHomeDir(); homeErr != nil {
-		return "", fmt.Errorf("failed to locate home directory; %w", homeErr)
-	} else {
-		return filepath.Join(homeDir, "meta"), nil
 	}
 }
